+++ optimum/neuron/utils/__init__.py
@@ -40,12 +40,12 @@ from .misc import (
     check_if_weights_replacable,
     get_stable_diffusion_configs,
     is_main_worker,
+    is_precompilation,
     replace_weights,
 )
 from .optimization_utils import get_attention_scores_sd, get_attention_scores_sdxl
 from .patching import DynamicPatch, ModelPatcher, Patcher, patch_everywhere, patch_within_function
 from .training_utils import (
     is_model_officially_supported,
-    is_precompilation,
     patch_transformers_for_neuron_sdk,
 )
